# web/openapi.json
{
  "components": {
    "schemas": {
...
      "AttendanceCourseRuns": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AttendanceCourseRun"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "AttendanceCourseRuns",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MakeupCreditBody": {
        "additionalProperties": false,
        "properties": {
          "makeup_session_unit_id": {
            "exclusiveMinimum": 0.0,
            "title": "Makeup Session Unit Id",
            "type": "integer"
          },
          "reason": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "makeup_session_unit_id",
          "reason"
        ],
        "title": "MakeupCreditBody",
        "type": "object"
      },
      "MakeupCreditResult": {
        "properties": {
          "attendance_id": {
            "title": "Attendance Id",
            "type": "integer"
          },
          "credited_status": {
            "const": "Present",
            "title": "Credited Status",
            "type": "string"
          },
          "denominator_units_added": {
            "const": 0,
            "title": "Denominator Units Added",
            "type": "integer"
          },
          "makeup_for_attendance_id": {
            "title": "Makeup For Attendance Id",
            "type": "integer"
          }
        },
        "required": [
          "attendance_id",
          "makeup_for_attendance_id",
          "credited_status",
          "denominator_units_added"
        ],
        "title": "MakeupCreditResult",
        "type": "object"
      },
      "MakeupOptions": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MakeupAbsenceOption"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "MakeupOptions",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "English Class API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/attendance/course-runs": {
      "get": {
        "operationId": "attendance_run_list_api_attendance_course_runs_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "english_class_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "English Class Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceCourseRuns"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Attendance Run List"
      }
    },
    "/api/attendance/makeup-options": {
      "get": {
        "operationId": "attendance_makeup_options_api_attendance_makeup_options_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "english_class_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "English Class Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MakeupOptions"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Attendance Makeup Options"
      }
    },
    "/api/attendance/{attendance_id}/makeup-credit": {
      "post": {
        "operationId": "attendance_makeup_credit_api_attendance__attendance_id__makeup_credit_post",
        "parameters": [
          {
            "in": "path",
            "name": "attendance_id",
            "required": true,
            "schema": {
              "title": "Attendance Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "english_class_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "English Class Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MakeupCreditBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MakeupCreditResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Attendance Makeup Credit"
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login"
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "logout_api_auth_logout_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "english_class_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "English Class Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout"
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "me_api_auth_me_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "english_class_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "English Class Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Me"
      }
    },
    "/api/course-runs/{course_run_id}/attendance-sessions": {
      "post": {
        "operationId": "attendance_session_create_api_course_runs__course_run_id__attendance_sessions_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_run_id",
            "required": true,
            "schema": {
              "title": "Course Run Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "english_class_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "English Class Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttendanceSessionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceSessionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Attendance Session Create"
      }
    },
    "/api/course-runs/{course_run_id}/session-units": {
      "get": {
        "operationId": "attendance_unit_list_api_course_runs__course_run_id__session_units_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_run_id",
            "required": true,
            "schema": {
              "title": "Course Run Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "english_class_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "English Class Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceSessionUnits"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Attendance Unit List"
      }
    },
...
    "/api/dashboard": {
      "get": {
        "operationId": "dashboard_api_dashboard_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "english_class_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "English Class Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dashboard"
      }
    },
...
  }
}